Move App into the modules folder
