+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import { App } from './App';
+import { App } from './modules/App/index';
 import './styles.css';
 
 const rootDiv = document.createElement('div');
